+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2024-12-19
+
 ### Added
-- Variable page size for the concurrent implementation
+- A benchmark for concurrent implementations
 
 ### Changed
+- A new **much faster** partially lock free concurrent implementation 
+- Reorganized the library
 
 ## [0.5.0] - 2024-09-18